Move APIs to common/
