+++ astra-sim/common/AstraComputeAPI.cc
@@ -3,7 +3,7 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "astra-sim/system/AstraComputeAPI.hh"
+#include "astra-sim/common/AstraComputeAPI.hh"
 
 using namespace AstraSim;
