--- astra-sim/system/AstraComputeAPI.cc
@@ -1,1 +1,0 @@
-../common/AstraComputeAPI.cc
