--- astra-sim/system/AstraComputeAPI.hh
@@ -1,1 +1,0 @@
-../common/AstraComputeAPI.hh
